==> 2. Feature Engineering/1-3. BOW Generation.ipynb <==
## Import

# for read data
import os

# for feature generation
import pandas as pd
import numpy as np
import warnings; warnings.filterwarnings("ignore")

# for feature decompose
from sklearn.decomposition import PCA 

### Read Data

path = (os.path.abspath("./input"))

train = pd.read_csv(path + '/X_train_DC.csv', parse_dates=['SALES_DATE', 'SALES_DAY'], encoding='cp949')
test = pd.read_csv(path + '/X_test_DC.csv', parse_dates=['SALES_DATE', 'SALES_DAY'], encoding='cp949')
y_train = pd.read_csv(path + '/y_train.csv').group

# Concat to One Data
data = pd.concat([train, test]).reset_index(drop=True)

# 추후 custid를 기준으로 train data로 만든 feature, test data로 만든 feature로 분리한다.
train_ID = train.custid.unique()
test_ID = test.custid.unique()

feature = pd.DataFrame({'custid':train.custid.unique().tolist() + test.custid.unique().tolist()})
feature

## Bag of Words
Category feature를 자체적으로 처리하는 Catboost에만 적용할 feature를 생성한다.

- **[goodcd]**<br>
  고객의 구매한 적이 있는 상품이면 1, 구매한 적이 없는 상품이면 0으로 나타낸 데이터이다.

OH_good =  pd.pivot_table(data.drop_duplicates(['custid','goodcd']), 
                          index='custid', columns='goodcd', values='tot_amt', aggfunc=np.size, fill_value=0).reset_index()
OH_good.columns = ['custid']+[f'상품_{i}' for i in OH_good.columns[1:]]

feature = pd.merge(feature, OH_good, on='custid', how='outer')
feature

- **[brd_nm]**<br>
  고객의 구매한 적이 있는 브랜드이면 1, 구매한 적이 없는 브랜드이면 0으로 나타낸 데이터이다.

OH_brd =  pd.pivot_table(data.drop_duplicates(['custid','brd_nm']),
                         index='custid', columns='brd_nm', values='tot_amt', aggfunc=np.size, fill_value=0).reset_index()
OH_brd.columns = ['custid']+[f'브랜드_{i}' for i in OH_brd.columns[1:]]

feature = pd.merge(feature, OH_brd, on='custid', how='outer')
feature

- **[corner_nm]**<br>
  고객의 구매한 적이 있는 코너이면 1, 구매한 적이 없는 코너이면 0으로 나타낸 데이터이다.

OH_cor =  pd.pivot_table(data.drop_duplicates(['custid','corner_nm']), 
                         index='custid', columns='corner_nm', values='tot_amt', aggfunc=np.size, fill_value=0).reset_index()
OH_cor.columns = ['custid']+[f'코너_{i}' for i in OH_cor.columns[1:]]

feature = pd.merge(feature, OH_cor, on='custid', how='outer')
feature

- **[pc_nm]**<br>
  고객의 구매한 적이 있는 상품군이면 1, 구매한 적이 없는 상품군이면 0으로 나타낸 데이터이다.

OH_pc =  pd.pivot_table(data.drop_duplicates(['custid','pc_nm']), 
                        index='custid', columns='pc_nm', values='tot_amt', aggfunc=np.size, fill_value=0).reset_index()
OH_pc.columns = ['custid']+[f'상품군_{i}' for i in OH_pc.columns[1:]]

feature = pd.merge(feature, OH_pc, on='custid', how='outer')
feature

- **[part_nm]**<br>
  고객의 구매한 적이 있는 파트이면 1, 구매한 적이 없는 파트이면 0으로 나타낸 데이터이다.

OH_part =  pd.pivot_table(data.drop_duplicates(['custid','part_nm']), 
                          index='custid', columns='part_nm', values='tot_amt', aggfunc=np.size, fill_value=0).reset_index()
OH_part.columns = ['custid']+[f'파트_{i}' for i in OH_part.columns[1:]]

feature = pd.merge(feature, OH_part, on='custid', how='outer')
feature

- **[team_nm]**<br>
  고객의 구매를 담당한 팀이면 1, 담당하지 않은 팀이면 0으로 나타낸 데이터이다.

OH_team =  pd.pivot_table(data.drop_duplicates(['custid','team_nm']), 
                          index='custid', columns='team_nm', values='tot_amt', aggfunc=np.size, fill_value=0).reset_index()
OH_team.columns = ['custid']+[f'팀_{i}' for i in OH_team.columns[1:]]

feature = pd.merge(feature, OH_team, on='custid', how='outer')
feature

- **[buyer_nm]**<br>
  고객의 구매한 상품의 바이어이면 1, 구매한 상품의 바이어가 아니면 0으로 나타낸 데이터이다.

OH_buyer =  pd.pivot_table(data.drop_duplicates(['custid','buyer_nm']), 
                           index='custid', columns='buyer_nm', values='tot_amt', aggfunc=np.size, fill_value=0).reset_index()
OH_buyer.columns = ['custid']+[f'바이어_{i}' for i in OH_buyer.columns[1:]]

feature = pd.merge(feature, OH_buyer, on='custid', how='outer')
feature

## PCA
Spare함으로 PCA를 통해 일부만을 저장한다.

# split data
feature_train = feature.query('custid in @train_ID').reset_index(drop=True)
feature_test = feature.query('custid in @test_ID').reset_index(drop=True)

# del custid
del feature_train['custid'], feature_test['custid']

max_d = num_d = feature_train.shape[1]

pca = PCA(n_components=max_d, random_state=0).fit(feature_train)
cumsum = np.cumsum(pca.explained_variance_ratio_) # 분산의 설명량을 누적합
num_d = np.argmax(cumsum >= 0.99) + 1             # 분산의 설명량이 99%이상 되는 차원의 수
if num_d == 1: num_d = max_d
    
pca = PCA(n_components=num_d, random_state=0)  
feature_train = pca.fit_transform(feature_train)
feature_test = pca.transform(feature_test)
print(feature_train.shape)

### Make train, test data

feature_train = pd.concat([pd.Series(train_ID, name='custid'), pd.DataFrame(feature_train)], axis=1)
feature_test = pd.concat([pd.Series(test_ID, name='custid'), pd.DataFrame(feature_test)], axis=1)

### Save train, test data

feature_train.to_csv(path + '/feature_train_BOW.csv', index=False, encoding='cp949')
feature_test.to_csv(path + '/feature_test_BOW.csv', index=False, encoding='cp949')
