==> 2. Feature Engineering/3. Selection.ipynb <==
## Imports

# for read data
import os
import pandas as pd
import warnings; warnings.filterwarnings("ignore")

# feature selection
from sklearn.model_selection import train_test_split
from catboost import CatBoostClassifier 
from lightgbm import LGBMClassifier 

# validation
from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import log_loss, make_scorer

## Exclude Low Importance Features
모든 Feature를 집어넣어 학습시킨 뒤 Feature Importance가 낮은 Feature는 제외하고 Selection을 진행한다.

- **[Catboost]**

path = (os.path.abspath("./input"))

feature_train1 = pd.read_csv(path +'/feature_train_transformation.csv', encoding='cp949')
feature_train2 = pd.read_csv(path +'/feature_train_W2V.csv', encoding='cp949')
feature_train3 = pd.read_csv(path +'/feature_train_BOW.csv', encoding='cp949')
feature_train = pd.merge(feature_train1, feature_train2, on='custid', how='outer')
feature_train = pd.merge(feature_train, feature_train3, on='custid', how='outer')

feature_test1 = pd.read_csv(path +'/feature_test_transformation.csv', encoding='cp949')
feature_test2 = pd.read_csv(path +'/feature_test_W2V.csv', encoding='cp949')
feature_test3 = pd.read_csv(path +'/feature_test_BOW.csv', encoding='cp949')
feature_test = pd.merge(feature_test1, feature_test2, on='custid', how='outer')
feature_test = pd.merge(feature_test, feature_test3, on='custid', how='outer')

y_train = pd.read_csv(path +'/y_train.csv', encoding='cp949').group

del feature_train['custid'], feature_test['custid']

X_train, X_dev, y__train, y_dev = train_test_split(feature_train, y_train, test_size=0.3, random_state=0)

# learning_rate만 지정한다.
model=CatBoostClassifier(iterations=100, learning_rate=0.03, objective='MultiClass', bootstrap_type ='Bayesian',
                           devices='0:1', task_type="GPU", random_state=0)

model.fit(X_train, y__train)

# 전체 Feature 사용: 1.5727539589558865
# 상위 171개 사용:  1.572028648061302, redefine_1 결과
# 상위 120개 사용:  1.5705520554742651, redefine_2 결과, STOP!
# 상위 104개 사용:  1.5726428439180589, redefine_3 결과
print(f'상위 {feature_train.shape[1]}개 사용: ',log_loss(y_dev, model.predict_proba(X_dev)))

# Feature Importance 확인
FI = pd.DataFrame({'Feature_Importances':model.feature_importances_})
display(FI.value_counts().sort_index())

# redefine features
# Log loss가 올라가는 importance 이면 그 사이값 재탐색 후 결정한다.
# limit = round(FI.value_counts().sort_index().index[1][0], 6)
# redefine_1 = FI.query('Feature_Importances > @limit').index
# redefine_2 = FI.query('Feature_Importances > @limit').index
# redefine_3 = FI.query('Feature_Importances > @limit').index

# feature_train = feature_train.iloc[:, redefine_3]
# feature_test = feature_test.iloc[:, redefine_3]

- **[LGBM]**

feature_train0 = pd.read_csv(path +'/feature_train_transformation_0.csv', encoding='cp949')
feature_train3 = pd.read_csv(path +'/feature_train_W2V.csv', encoding='cp949')
feature_train4 = pd.read_csv(path +'/feature_train_BOW.csv', encoding='cp949')
feature_train = pd.merge(feature_train0, feature_train3, on='custid', how='outer')
feature_train = pd.merge(feature_train, feature_train4, on='custid', how='outer')

feature_test0 = pd.read_csv(path +'/feature_test_transformation_0.csv', encoding='cp949')
feature_test3 = pd.read_csv(path +'/feature_test_W2V.csv', encoding='cp949')
feature_test4 = pd.read_csv(path +'/feature_test_BOW.csv', encoding='cp949')
feature_test = pd.merge(feature_test0, feature_test3, on='custid', how='outer')
feature_test = pd.merge(feature_test, feature_test4, on='custid', how='outer')

y_train = pd.read_csv(path +'/y_train.csv', encoding='cp949').group

del feature_train['custid'], feature_test['custid']

X_train, X_dev, y__train, y_dev = train_test_split(feature_train, y_train, test_size=0.3, random_state=0)

# learning_rate만 지정한다.
model=LGBMClassifier(learning_rate=0.03,
                     objective='multiclass', metrics='multi_logloss', num_gpu=1, random_state=0)

model.fit(X_train, y__train)

# 전체 Feature 사용: 1.5479148067922546
# 상위 4656개 사용:  1.5472279449561581, redefine_1 결과, STOP!
# 상위 4021개 사용:  1.548362171129472, redefine_2 결과

print(f'상위 {feature_train.shape[1]}개 사용: ',log_loss(y_dev, model.predict_proba(X_dev)))

# Feature Importance 확인
FI = pd.DataFrame({'Feature_Importances':model.feature_importances_})
display(FI.value_counts().sort_index())

# redefine features
# Log loss가 올라가는 importance 직전에서 멈춘다.
# REDEFINE_1 = FI.query('Feature_Importances > 1').index 
# REDEFINE_2 = FI.query('Feature_Importances > 1').index 

# feature_train = feature_train.iloc[:, REDEFINE_2]
# feature_test = feature_test.iloc[:, REDEFINE_2]

### $~~~~$ Save data

feature_train1 = pd.read_csv(path +'/feature_train_transformation.csv', encoding='cp949')
feature_train2 = pd.read_csv(path +'/feature_train_W2V.csv', encoding='cp949')
feature_train3 = pd.read_csv(path +'/feature_train_BOW.csv', encoding='cp949')
feature_train = pd.merge(feature_train1, feature_train2, on='custid', how='outer')
feature_train = pd.merge(feature_train, feature_train3, on='custid', how='outer')

feature_test1 = pd.read_csv(path +'/feature_test_transformation.csv', encoding='cp949')
feature_test2 = pd.read_csv(path +'/feature_test_W2V.csv', encoding='cp949')
feature_test3 = pd.read_csv(path +'/feature_test_BOW.csv', encoding='cp949')
feature_test = pd.merge(feature_test1, feature_test2, on='custid', how='outer')
feature_test = pd.merge(feature_test, feature_test3, on='custid', how='outer')

train_ID, test_ID = feature_train.custid, feature_test.custid
del feature_train['custid'], feature_test['custid']

# Catboost
# logloss: 1.5705520554742651
cat_train = pd.concat([train_ID, 
                       feature_train.iloc[:, redefine_1].iloc[:, redefine_2]], axis=1)
cat_train.to_csv(path+'/CAT_train.csv', index=False, encoding='cp949')

cat_test = pd.concat([test_ID, 
                      feature_test.iloc[:, redefine_1].iloc[:, redefine_2]], axis=1)
cat_test.to_csv(path+'/CAT_test.csv', index=False, encoding='cp949')

# lgbmboost
# logloss: 1.5472279449561581
lgbm_train = pd.concat([train_ID, feature_train.iloc[:, REDEFINE_1]], axis=1)
lgbm_train.to_csv(path+'/LGBM_train.csv', index=False, encoding='cp949')

lgbm_test = pd.concat([test_ID, feature_test.iloc[:, REDEFINE_1]], axis=1)
lgbm_test.to_csv(path+'/LGBM_test.csv', index=False, encoding='cp949')
